--- pkg/generic/thrift/util_test.go
@@ -1,41 +1,0 @@
-/*
- * Copyright 2024 CloudWeGo Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package thrift
-
-import (
-	"testing"
-
-	"github.com/cloudwego/kitex/internal/test"
-)
-
-func TestSplitType(t *testing.T) {
-	pkg, name := splitType(".A")
-	test.Assert(t, pkg == "")
-	test.Assert(t, name == "A")
-
-	pkg, name = splitType("foo.bar.A")
-	test.Assert(t, pkg == "foo.bar")
-	test.Assert(t, name == "A")
-
-	pkg, name = splitType("A")
-	test.Assert(t, pkg == "")
-	test.Assert(t, name == "A")
-
-	pkg, name = splitType("")
-	test.Assert(t, pkg == "")
-	test.Assert(t, name == "")
-}
